flask/requirements.txt
boto3 
botocore
flask
python-decouple
requests
